% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{as}
\alias{as}
\title{Convert \code{seurat} to \code{SingleCellExperiment} objects}
\arguments{
\item{from}{A \code{\linkS4class{seurat}} object.}
}
\value{
A \code{\linkS4class{SingleCellExperiment}} object.
}
\description{
Convert \code{seurat} to \code{SingleCellExperiment} objects
}
\examples{
# Example data ----

stopifnot(require(Seurat))
pbmc_raw <- read.table(
  file = system.file('extdata', 'pbmc_raw.txt', package = 'Seurat'),
  as.is = TRUE
)
pbmc_small <- CreateSeuratObject(raw.data = pbmc_raw)
pbmc_small

# Example ----

as(pbmc_small, "SingleCellExperiment")

}
